requirements.txt
Mastodon.py
python-dotenv
